requirements.txt
Scrapy==0.24.4
requests=2.4.3
django-bootstrap3==4.11.0
"django<1.7"
python-dateutil==2.2
fake-factory==0.4.2
